requirements.txt
Flask>=1.0
opencv-python>=3.4
pyserial>=3.0
requests>=2.0